Fix checkpoints downloads
This commit replaces the checkpoint download URLs since they changed
after the first alpha release name was updated.

Also, I improve the error handling when downloading checkpoints if it
happens that there's an issue downloading them.

Finally, I added the option to not-prompt for download if the user is
not running terran inside a interactive terminal.